feat: add NLP-enhanced retrieval modes to LongMemEval benchmark
Add nlp_aaak and nlp_hybrid modes to longmemeval_bench.py that use NLP
providers for entity detection, sentence splitting, and compression during
ingestion. Rewrite bench_nlp_providers.py as a thin wrapper that runs
baseline vs NLP-enhanced comparisons on the LongMemEval dataset, producing
directly comparable Recall@k and NDCG@k scores.
